wallet mod disk to file
